+++ inc/HM11/hm11.h
@@ -16,8 +16,8 @@
 #define BAUD_VALUE_9600      650 // 650.04
 #define APRS_BUFF_SIZE       256
 
-#define UART_TX_BUFF_SIZE 20
-#define UART_RX_BUFF_SIZE 8
+#define BLE_TX_BUFF_SIZE 20
+#define BLE_RX_BUFF_SIZE 8
 
 /* ----- Packet Parameters ----- */
 #define HEADER_1    0x41